feat(bulk-move): fall back to loading the English forum name
If the forum name is not available in the user's display language, now
the English forum name will be shown as a fall back if available.

This is made to replicate the logic used by the Community Console to
choose the default language when preloading the forum data.

Fixed: twpowertools:243
Bug: twpowertools:51
Change-Id: I6a6a69642c53ea308bee2198f18494e1756739d2