fix(xhr-interceptor): ensure XHR interceptor runs before anything else
We used to inject the XHR interceptor script from an isolated content
script to the main world. In firefox, the XHR interceptor ran after the
Community Console was loaded (and already made some requests) due to the
way that Firefox treats scripts added dynamically to the page.

This CL fixes this by running the XHR interceptor directly as a content
script in the main world.

Fixed: twpowertools:232
Change-Id: I6a6a69644b0241475a7cfa3152bad397b3b37173