fix(generic-action-button): prevent creating them disabled
This CL fixes the fact that sometimes generic thread list actions were
created in a disabled state (even if they worked when clicked).

This was due to the fact that the generic action button injector cloned
the "mark as duplicate" button, which was sometimes marked as disabled
depending on the selected threads. We fix this by instead cloning the
"mark as (not) read" buttons if possible, since those are never disabled.

This was the behavior found in the batch lock button before commit
ece937d5b4d0f445c30cf11cc2e4db285f431c60, which introduced the
regression for that feature.

Fixed: twpowertools:241
Change-Id: I6a6a6964ca388da379b6203a20faa348e19f6b07