feat: sync some translations from the Community Console
Some strings we use in our project are identical to some already
existing in the Community Console. To ease the work of translators, this
CL allows importing existing translations in the Community Console
(which are simple enough so we don't have to worry about copyright).

Some strings have been added to the sync configuration, and have been
already imported in this CC. Syncronization will occur manually.

Change-Id: I6a6a6964de35acb064ba64f346f7f0bd7c77d662