fix: ignore CC requests used to add filter counters to the drawer
The Community Console has changed and a feature was apparently added to
show counters next to some threads in the drawer. This feature loads the
list of threads of each filter via a ViewForum request, and then counts
the number of threads by counting how many there are in the response.

This causes some bugs in our extension because we are treating these
requests as if they were made to display a thread list.

This CL fixes all these issues by ignoring these requests.

Fixed: twpowertools:246
Change-Id: I6a6a6964a4091c818a5cf6339ff5cae680ea32b4