chore: remove code related to optional permissions
When we wanted to support running the extension in sandbox domains
(support-*.sandbox.google.com), we introduced the ability to associate
features with required optional permissions that should be enabled by
the user.

This code hasn't been used in the end, since sandbox domains were shut
down. Thus, this CL removes this code.

Bug: twpowertools:247
Change-Id: I6a6a6964f715e92a2d100f957f1be73203daa50e