fix(options): stop showing feature as disabled when remotely disabled
When a kill switch was active (AKA, the feature was remotely disabled),
when a feature was enabled by a user, it would be shown as disabled
instead.

This CL fixes this by letting the user choose whether they want the
feature enabled or disabled independently of the kill switch.

A follow-up CL will implement showing the kill switch status.

Bug: twpowertools:63
Change-Id: I6a6a696499da8b514a2ffabe2700572072f221e4