feat(options): enable new options page for anglophone users
The new options page will be dogfooded with English-speaking users.
While the dogfood is underway, localization will be made to other
languages.

After the translation of the new options page to a new language is
complete, this language will be added to the supported languages list.
Before this, users for these languages will see the old options page.

Bug: twpowertools:63
Change-Id: I6a6a6964537191ead2d696025072afa2a010a9f4