feat(options): show whether features have an active kill switch
This CL adds a hint text that appears when a feature has an active kill
switch (i.e., it has been remotely disabled) in the feature card. This
lets users know what features have been remotely disabled, like the old
options page did.

It also changes the checkbox color to the error color when a kill switch
is active.

Bug: twpowertools:63
Change-Id: I6a6a6964e4cc880ff1dd2f533fc61d27abf11944