fix(options-provider): make isEnabled work again with internal options
When calling |isEnabled()| with a boolean internal option, the response
would always be |false| due to a regression caused in
82229aef17df05becc275796d5a97497a9747805.

In that commit, an additional check was added so the method would always
return |false| for options which didn't have |killSwitchType| set to
|KillSwitchType.Option| in the options prototype. This was a check
introduced because we believed we only used that method for feature
options, but that is not the case and thus caused this bug. We now allow
other boolean options to use that method as well.

To improve type safety, we should mark in the options prototype the type
of each option (boolean, string, ...) and only allow calls to this
method for boolean options.

Bug: twpowertools:247
Change-Id: I6a6a696469db2136da02166320e8656ac50800c9