refactor: remove workaround for features not available in some browsers
For features that are not available in all browsers, we have to work
around the fact that |OptionCodename| didn't include the codenames for
these options, but ts-loader insists on trying to compile the logic of
non-available options, for which we assume that the option codename is
part of |OptionCodename|.

This CL improves this by always adding these options to
|rawOptionConfigs|. The option codenames will automatically be part of
|OptionCodename|, but the features won't be shown in the options page
nor will work, since the code is removed from the build. However, this
change has the side effect that the options will be set in the extension
storage, though as we said they will be unused.

Change-Id: I6a6a6964fdfe83b86abf7592840fe0e7edc1688e