feat: add MWOptionsConfigurationRepositoryAdapter
This adapter will let us access options configuration from Main World
scripts, by sending requests through a bridge to a Content Script server
(introduced in this CL as well), which will request the actual options
configuration via the existing repository adapter.

This is in preparation for a new feature which will use OptionsProvider
to access the configuration from a Main World script.

Change-Id: I6a6a696491acccce266743314a0c8c89bd30c561