build: split js compilation from common_ts_project
Combining Typescript and Javascript in builds is problematic, so each
package will build Javascript in their own way, with the help of the new
common_js_project macro.

This CL also refactors the build of //src/common/options to use the new
macro, and //:bundle.bzl so that it also includes transitive types.

Bug: twpowertools:255
Change-Id: Id64237e21c20d9ece14ee2584146a1756a6a6964