build: add //src:non_compiled_source target
The //src:non_packaged_source target contains non-packaged source, but
we also need to include packaged non-compiled source code to
pass through to Webpack and Vitest in the root of the project.

This is useful because when migrating all source code to individual
Bazel packages, we will sometimes have subpackages that will not be able
to be compiled because their dependencies are not yet a Bazel package.
The source code of these folders will initially be passed to the
//src:non_compiled_source target along with the
//src:non_packaged_source target.

Bug: twpowertools:255
Change-Id: I20cd3214f84f7a72700b435de78beff26a6a6964