build: add package groups for the different layers
This CL adds some initial definitions for package groups that define the
packages that belong to each different layer. This will let us define
the visibility of packages easily, by stating that each package should
be available to a certain set of layers.

As we add more Bazel packages, we may add them to the groups so that
they can depend on other packages which set visibility in the
aforementioned manner.

Bug: twpowertools:255
Change-Id: I160797ed37003b5fafb3672546403ad16a6a6964