chore: add renovate-bot as owner of some bazel files
Renovate will update the MODULE.bazel and lock files, as well as the
Bazel version when there are new updates. Also, when bazelrc-preset.bzl
is updated, it will also regenerate the tools/preset.bazelrc file.

This CL allows Renovate to auto-merge changes to these files without
further approval from other maintainers if we decide to enable this in
the future.

Bug: twpowertools:255
Change-Id: I672b2642d913529726ca9ca8262065496a6a6964