build: fix running printOptionsPrototypeJSON
When running the binary via Bazel, it would complain with an
ERR_MODULE_NOT_FOUND error:

  Error [ERR_MODULE_NOT_FOUND]: Cannot find module
  '.../tools/printOptionsPrototypeJSON/printOptionsPrototypeJSON_/printOptionsPrototypeJSON.runfiles/_main/src/common/options/optionsPrototype'
  imported from
  .../tools/printOptionsPrototypeJSON/printOptionsPrototypeJSON_/printOptionsPrototypeJSON.runfiles/_main/tools/printOptionsPrototypeJSON/printOptionsPrototypeJSON.js

This CL fixes this by changing tsconfig.json for library packages and
the printOptionsPrototypeJSON package so that `module = commonjs` and
`moduleResolution = node10`. Before this change, `moduleResolution` was
`bundle`, which was fine when the files were bundled by Webpack, but not
when using the compiled Typescript directly as is the case in
printOptionsPrototypeJSON.

This change is also compatible with the main build that uses ts-loader
in Webpack.

Bug: twpowertools:255
Change-Id: Id10fe8e543bbd20a324e67f9ae577e9d6a6a6964