chore(deps): automerge all updates to renovate
The renovate package is only used in //:renovate_config_test, which
makes sure that our renovate.json file is correct. Thus, it is safe to
update it to any version automatically if the tests pass.

New major versions may have breaking changes, but if there are any, our
config will break.

Change-Id: Id2aaf11b9435e61ce89a9ba32976d3006a6a6964