build: refactor skipping "web-ext lint" test
Now the test is skipped using the |target_compatible_with| attribute,
which more accurately reflects the status of the test as "skipped" when
running Bazel with an incompatible configuration.

The "maybe_" prefix has also been removed from the target name since now
it is clear whether the test is skipped or not.

Bug: twpowertools:255
Change-Id: I27fc8ef79e3a675d50a4a0614a33ae096a6a6964