fix(cc-dark-theme): fix dark theme in non-bazel builds
In the default non-Bazel builds of the extension, the stylesheet was
named main.bundle.css and not main.css (as I incorrectly expected) like
in Bazel builds. This caused the extension to try to load a non-existing
stylesheet (main.css) and the dark theme didn't work at all.

This CL unifies the file name independently of which build system is
used, so the feature continues to work in all builds.

This was noticed in the Canary release by Brett:
https://groups.google.com/g/twpowertools-discuss/c/8oNZJaqNkxI

Change-Id: I7c088af83eeec606a206149236c268b06a6a6964