build: prepare for source map support in bazel builds
In Bazel builds, we don't have support for extracting the source maps
generated by ts_project. This CL lays the ground for supporting this in
the future.

Unfortunately, this is incompatible with preprocessorLoader, but once we
use our own preprocessor and remove that loader, source map extraction
will work properly.

Bug: twpowertools:255
Change-Id: Id98ffa47de770768ffa508e2bc0d1fee6a6a6964