build: stop making most packages visible to //src:__pkg__
Most packages were visible to //src:__pkg__ (or //:__pkg__) because
uncompiled code (with regards to Bazel) passed to Webpack also needed
access/visibility to compiled source code while we were transitioning to
Bazel.

Now this is no longer the case because all JS/TS source code is compiled
with Bazel.

Bug: twpowertools:256
Change-Id: I27870cc9a0103566d5d35020f89c09cf6a6a6964