chore: remove todos we won't act on
When I started the Bazel migration, I wrote some TODOs to move the
common console stylesheet out of the static folder. However, we should
in fact relocate all files under static to more appropriate places.

Thus, we're removing the TODOs until sometime in the future I maybe do
this bigger refactor.

This concludes the cleaning we had to do after the Bazel migration, so
this commit marks twpowertools:256 as fixed. However, more unknown
cleanup might be required. That work will also be associated to that
bug, even if closed.

Fixed: twpowertools:256
Change-Id: I192dc873a3902242340157b5e980d29c6a6a6964