fix: show kill switch badge after reenabling the extension
When the extension was disabled and reenabled with an active kill
switch, the badge warning about the kill switch would not appear in the
extension icon.

This CL fixes this issue by ensuring we update the badge each time the
service worker starts, so that it will be updated when the extension is
reenabled.

Ideally, this update would only run after the extension is reenabled,
but I haven't figured out a way to do that.

Change-Id: I176c1f7bcb8d634a35595b1450cb4a2a6a6a6964