chore: make LSPs find files generated by Bazel
When developing locally, the editor's LSP will often complain about
missing files--those that are generated by Bazel and are not committed
to the source code tree.

This CL improves the developer experience by adding the bazel-bin as a
root directory for Typescript in the tsconfig.json file (the file that
is used by default by LSPs), while moving the former tsconfig.json to a
file named tsconfig.base.json.

The current tsconfig.json will only be used by LSPs, while
tsconfig.base.json will continue to be used to build the project.

Change-Id: I5c09efca81073f3a8c6967613029ef1c6a6a6964