refactor(dependencies-provider): create classes in the composition root
The classes used to be created in the dependencies provider when the
dependency was asked to be retrieved but it was not yet created.
However, some dependencies will need to have dependencies injected via
their constructor.

Thus, this CL moves the creation back to the composition root, so it can
inject the dependencies as needed.

This fixes the control freak antipattern that was present in the
dependencies provider.

Change-Id: I1aacb755157e3f5ce0171445063f07dd6a6a6964