refactor(xhr-interceptor): use OptionsProviderPort
The XHR interceptor is the last place that used the
PartialOptionsWatcher. This CL substitutes it with the use of
OptionsProviderPort via the MWOptionsConfigurationRepositoryAdapter.

This will make the feature slightly less performant, but aids in
maintainability. Gaining the same performance we had before will be
tracked at https://iavm.xyz/b/twpowertools/260.

We will clean up the PartialOptionsWatcher in a follow-up commit.

Bug: twpowertools:254
Change-Id: I032b29e63cf2496b54669eed51d9af136a6a6964