chore: add foundations for sync storage area migrations
This is the first CL in a series which will add support for migrating
the sync storage area when its schema (i.e., its form) needs to be
changed.

Here we introduce the StorageAreaMigration entity, a package in which
to save migrations, and the schema which represents the current state of
the sync storage area.

Bug: twpowertools:261
Change-Id: Idfebb67a0ad67628bf196d8efdeeb1156a6a6964